Add VS2019 Support

+++ src/game/client/deferred/vgui/vgui_editor_root.cpp
@@ -615,10 +615,10 @@ void CVGUILightEditor::OnCursorMoved_Internal( int &x, int &y )
 				vecDelta[1].NormalizeInPlace();
 
 				float flAngle[2] = {
-					atan2( DotProduct( vDirections[ iDirIndices[iAxis][0] ], vecDelta[0] ),
-						DotProduct( vDirections[ iDirIndices[iAxis][1] ], vecDelta[0] ) ),
-					atan2( DotProduct( vDirections[ iDirIndices[iAxis][0] ], vecDelta[1] ),
-						DotProduct( vDirections[ iDirIndices[iAxis][1] ], vecDelta[1] ) ),
+					static_cast<float>(atan2( DotProduct( vDirections[ iDirIndices[iAxis][0] ], vecDelta[0] ),
+						DotProduct( vDirections[ iDirIndices[iAxis][1] ], vecDelta[0] ) )),
+					static_cast<float>(atan2( DotProduct( vDirections[ iDirIndices[iAxis][0] ], vecDelta[1] ),
+						DotProduct( vDirections[ iDirIndices[iAxis][1] ], vecDelta[1] ) )),
 				};
 
 				float flRotate = flAngle[1] - flAngle[0];
