+++ src/game/server/ai_behavior_lead.cpp
@@ -1538,7 +1538,7 @@ void CAI_LeadGoal::InputActivate( inputdata_t &inputdata )
 	AI_LeadArgs_t leadArgs = { 
 		GetGoalEntityName(), 
 		STRING(m_iszWaitPointName), 
-		m_spawnflags, 
+		static_cast<unsigned int>(m_spawnflags), 
 		m_flWaitDistance, 
 		m_flLeadDistance, 
 		m_flRetrieveDistance,
